Remove all of the old tasks as what was important from them was used already.
