--- buildSrc/src/main/java/cc/squirreljme/plugin/tasks/test/EmulatedTestClassDescriptor.java
@@ -1,72 +1,0 @@
-// -*- Mode: Java; indent-tabs-mode: t; tab-width: 4 -*-
-// ---------------------------------------------------------------------------
-// Multi-Phasic Applications: SquirrelJME
-//     Copyright (C) Stephanie Gawroriski <xer@multiphasicapps.net>
-// ---------------------------------------------------------------------------
-// SquirrelJME is under the GNU General Public License v3+, or later.
-// See license.mkd for licensing and copyright information.
-// ---------------------------------------------------------------------------
-
-package cc.squirreljme.plugin.tasks.test;
-
-import org.gradle.api.internal.tasks.testing.DefaultTestClassDescriptor;
-import org.gradle.api.internal.tasks.testing.TestDescriptorInternal;
-
-/**
- * Describes a class being tested.
- *
- * @since 2020/03/06
- */
-public class EmulatedTestClassDescriptor
-	extends DefaultTestClassDescriptor
-{
-	/** The owning suite. */
-	protected final EmulatedTestSuiteDescriptor suite;
-	
-	/**
-	 * Initializes the class descriptor.
-	 *
-	 * @param __suite The owning suite.
-	 * @param __className The class being tested.
-	 * @since 2020/03/06
-	 */
-	public EmulatedTestClassDescriptor(EmulatedTestSuiteDescriptor __suite,
-		String __className)
-	{
-		super(new EmulatedTestId(), __className, __className);
-		
-		if (__suite == null)
-			throw new NullPointerException("No suite specified.");
-		
-		this.suite = __suite;
-	}
-
-	/**
-	 * {@inheritDoc}
-	 * @since 2020/03/06
-	 */
-    @Override
-    public String getDisplayName() {
-        return this.suite.getDisplayName() + ":" + this.getClassName();
-    }
-	
-	/**
-	 * {@inheritDoc}
-	 * @since 2020/03/06
-	 */
-    @Override
-	public Object getOwnerBuildOperationId()
-	{
-		// This must always be set otherwise Gradle will crash
-		return this.getId();
-	}
-	
-	/**
-	 * {@inheritDoc}
-	 * @since 2020/03/06
-	 */
-    @Override
-    public TestDescriptorInternal getParent() {
-        return this.suite;
-    }
-}
