--- buildSrc/src/main/java/cc/squirreljme/plugin/tasks/test/EmulatedTestId.java
@@ -1,19 +1,0 @@
-// -*- Mode: Java; indent-tabs-mode: t; tab-width: 4 -*-
-// ---------------------------------------------------------------------------
-// Multi-Phasic Applications: SquirrelJME
-//     Copyright (C) Stephanie Gawroriski <xer@multiphasicapps.net>
-// ---------------------------------------------------------------------------
-// SquirrelJME is under the GNU General Public License v3+, or later.
-// See license.mkd for licensing and copyright information.
-// ---------------------------------------------------------------------------
-
-package cc.squirreljme.plugin.tasks.test;
-
-/**
- * Emulated test ID.
- *
- * @since 2020/03/06
- */
-public final class EmulatedTestId
-{
-}
