--- buildSrc/src/main/java/cc/squirreljme/plugin/tasks/test/EmulatedTestUtilities.java
@@ -1,562 +1,0 @@
-// -*- Mode: Java; indent-tabs-mode: t; tab-width: 4 -*-
-// ---------------------------------------------------------------------------
-// Multi-Phasic Applications: SquirrelJME
-//     Copyright (C) Stephanie Gawroriski <xer@multiphasicapps.net>
-// ---------------------------------------------------------------------------
-// SquirrelJME is under the GNU General Public License v3+, or later.
-// See license.mkd for licensing and copyright information.
-// ---------------------------------------------------------------------------
-
-package cc.squirreljme.plugin.tasks.test;
-
-import cc.squirreljme.plugin.SquirrelJMEPluginConfiguration;
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Deque;
-import java.util.LinkedHashSet;
-import java.util.LinkedList;
-import java.util.Objects;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
-import org.gradle.api.Project;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.Dependency;
-import org.gradle.api.artifacts.ProjectDependency;
-import org.gradle.api.internal.tasks.testing.DefaultTestOutputEvent;
-import org.gradle.api.internal.tasks.testing.TestCompleteEvent;
-import org.gradle.api.internal.tasks.testing.TestDescriptorInternal;
-import org.gradle.api.internal.tasks.testing.TestStartEvent;
-import org.gradle.api.tasks.testing.TestOutputEvent;
-import org.gradle.api.tasks.testing.TestResult;
-import org.gradle.jvm.tasks.Jar;
-
-/**
- * Utilities for tests.
- *
- * @since 2020/03/06
- */
-public final class EmulatedTestUtilities
-{
-	/** Main configurations. */
-	private static final String[] _MAIN_CONFIGS =
-		new String[]{"api", "implementation"};
-	
-	/** Test configurations. */
-	private static final String[] _TEST_CONFIGS =
-		new String[]{"testApi", "testImplementation"};
-	
-	/** All configurations. */
-	private static final String[] _ALL_CONFIGS =
-		new String[]{"api", "implementation", "testApi", "testImplementation"};
-	
-	/** Normal JAR task. */
-	private static final String _MAIN_JAR_TASK =
-		"jar";
-	
-	/** Test JAR task. */
-	private static final String _TEST_JAR_TASK =
-		"testJar";
-	
-	/**
-	 * Not used.
-	 *
-	 * @since 2020/03/06
-	 */
-	private EmulatedTestUtilities()
-	{
-	}
-	
-	/**
-	 * Returns the test classpath of the given project.
-	 *
-	 * @param __project The project to get the classpath of.
-	 * @param __test Get the test class path?
-	 * @return The project classpath.
-	 * @since 2020/03/07
-	 */
-	public static Iterable<Path> classPath(Project __project, boolean __test)
-	{
-		// Which configuration to extract?
-		String[] configs = (__test ? EmulatedTestUtilities._TEST_CONFIGS :
-			EmulatedTestUtilities._MAIN_CONFIGS);
-		String inJar = (__test ? EmulatedTestUtilities._TEST_JAR_TASK :
-			EmulatedTestUtilities._MAIN_JAR_TASK);
-		
-		// Projects to evaluate and the projects we looked at
-		Deque<Project> eval = new LinkedList<>();
-		Set<Project> did = new TreeSet<>();
-		
-		// Start at our current project
-		eval.add(__project);
-		
-		// Process all projects for output
-		LinkedList<Path> rv = new LinkedList<>();
-		while (!eval.isEmpty())
-		{
-			// Only evaluate project once
-			Project at = eval.removeFirst();
-			if (did.contains(at))
-				continue;
-			did.add(at);
-			
-			// Only consider proper SquirrelJME projects
-			SquirrelJMEPluginConfiguration conf =
-				SquirrelJMEPluginConfiguration.configurationOrNull(at);
-			if (conf == null)
-				continue;
-			
-			// Add the JARs of execution, but for this project use the
-			// main/test JAR accordingly to what we want but for dependencies
-			// always use the main JAR because we do not want tests at all
-			Jar jarTask = (Jar)at.getTasks().findByName((at == __project ?
-				inJar : EmulatedTestUtilities._MAIN_JAR_TASK));
-			if (jarTask != null)
-				for (File file : jarTask.getOutputs().getFiles())
-					rv.addFirst(file.toPath().toAbsolutePath());
-			
-			// If this our own project, we need the main JAR as well, so make
-			// sure it is used!
-			if (at == __project)
-			{
-				Jar mainJar = (Jar)at.getTasks().findByName(
-					EmulatedTestUtilities._MAIN_JAR_TASK);
-				if (mainJar != null)
-					for (File file : mainJar.getOutputs().getFiles())
-						rv.addFirst(file.toPath().toAbsolutePath());
-			}
-			
-			// Process each configurations dependencies
-			for (String config : (at == __project ?
-				(__test ? EmulatedTestUtilities._ALL_CONFIGS: configs) :
-				EmulatedTestUtilities._MAIN_CONFIGS))
-			{
-				// If no configuration is set
-				Configuration buildConfig = at.getConfigurations()
-					.findByName(config);
-				if (buildConfig == null)
-					continue;
-					
-				// Process dependencies for project
-				for (Dependency dependency : buildConfig.getDependencies())
-				{
-					// Only consider projects
-					if (!(dependency instanceof ProjectDependency))
-						continue;
-					
-					// Put in the project to evaluate
-					eval.addLast(((ProjectDependency)dependency)
-						.getDependencyProject());
-				}
-			}
-		}
-		
-		return rv;
-	}
-	
-	/**
-	 * Returns the test classpath of the given project.
-	 *
-	 * @param __project The project to get the classpath of.
-	 * @param __test Get the test class path?
-	 * @return The project classpath.
-	 * @since 2020/03/07
-	 */
-	public static String classpathString(Project __project, boolean __test)
-	{
-		return EmulatedTestUtilities.classpathAsString(
-			EmulatedTestUtilities.classPath(__project, __test));
-	}
-	
-	/**
-	 * Returns the class path as a string.
-	 *
-	 * @param __paths Class paths.
-	 * @return The class path as a string.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2020/08/21
-	 */
-	public static String classpathAsString(Path... __paths)
-		throws NullPointerException
-	{
-		if (__paths == null)
-			throw new NullPointerException("NARG");
-		
-		return EmulatedTestUtilities.classpathAsString(Arrays.asList(__paths));
-	}
-	
-	/**
-	 * Returns the class path as a string.
-	 *
-	 * @param __paths Class paths.
-	 * @return The class path as a string.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2020/02/29
-	 */
-	public static String classpathAsString(Iterable<Path> __paths)
-		throws NullPointerException
-	{
-		if (__paths == null)
-			throw new NullPointerException("NARG");
-		
-		StringBuilder sb = new StringBuilder();
-		
-		for (Path path : __paths)
-		{
-			if (sb.length() > 0)
-				sb.append(File.pathSeparatorChar);
-			sb.append(path);
-		}
-		
-		return sb.toString();
-	}
-	
-	/**
-	 * Returns the emulator class path.
-	 *
-	 * @param __p The project for scanning.
-	 * @param __emu The emulator to base off.
-	 * @return The emulator class path.
-	 * @since 2020/03/07
-	 */
-	public static Iterable<Object> emulatorClassPath(Project __p, String __emu)
-	{
-		// Find emulator projects
-		Project rootProject = __p.getRootProject();
-		Project emuProject = rootProject.findProject(
-			":emulators:" + __emu + "-vm");
-		
-		Collection<Object> rv = new LinkedList<>();
-		
-		// Add the emulator run-time classpath along with the JAR it uses
-		// for execution
-		rv.add(emuProject.getConfigurations().getByName("runtimeClasspath")
-			.getFiles());
-		rv.add(((Jar)emuProject.getTasks().getByName("jar")).getOutputs()
-			.getFiles());
-		
-		return rv;
-	}
-	
-	/**
-	 * Fails at the specified time.
-	 * 
-	 * @param __millis The time to fail at.
-	 * @return The completion event.
-	 * @since 2020/06/22
-	 */
-	private static TestCompleteEvent failAt(long __millis)
-	{
-		return new TestCompleteEvent(__millis,
-			TestResult.ResultType.FAILURE);
-	}
-	
-	/**
-	 * Returns a failing completion.
-	 *
-	 * @return An event.
-	 * @since 2020/03/06
-	 */
-	public static TestCompleteEvent failNow()
-	{
-		return EmulatedTestUtilities.failAt(System.currentTimeMillis());
-	}
-	
-	/**
-	 * Outputs a message.
-	 *
-	 * @param __dest The destination.
-	 * @param __msg The message.
-	 * @return The message event.
-	 * @since 2020/03/06
-	 */
-	public static TestOutputEvent output(TestOutputEvent.Destination __dest,
-		String __msg)
-	{
-		return new DefaultTestOutputEvent(__dest, __msg);
-	}
-	
-	/**
-	 * Outputs a message.
-	 *
-	 * @param __msg The message.
-	 * @return The message event.
-	 * @since 2020/03/06
-	 */
-	public static TestOutputEvent output(String __msg)
-	{
-		return EmulatedTestUtilities.output(
-			TestOutputEvent.Destination.StdOut, __msg);
-	}
-	
-	/**
-	 * Outputs a message.
-	 *
-	 * @param __msg The message.
-	 * @return The message event.
-	 * @since 2020/03/06
-	 */
-	public static TestOutputEvent outputErr(String __msg)
-	{
-		return EmulatedTestUtilities.output(
-			TestOutputEvent.Destination.StdErr, __msg);
-	}
-	
-	/**
-	 * Passes at the given time.
-	 * 
-	 * @param __millis The time to pass at.
-	 * @return The completion event.
-	 * @since 2020/06/22
-	 */
-	private static TestCompleteEvent passAt(long __millis)
-	{
-		return new TestCompleteEvent(__millis,
-			TestResult.ResultType.SUCCESS);
-	}
-	
-	/**
-	 * Returns a passing completion.
-	 *
-	 * @return An event.
-	 * @since 2020/03/06
-	 */
-	public static TestCompleteEvent passNow()
-	{
-		return EmulatedTestUtilities.passAt(System.currentTimeMillis());
-	}
-	
-	/**
-	 * Either passes or fails.
-	 *
-	 * @param __pass Does the test pass?
-	 * @return The event.
-	 * @since 2020/03/06
-	 */
-	public static TestCompleteEvent passOrFailNow(boolean __pass)
-	{
-		return (__pass ? EmulatedTestUtilities.passNow() :
-			EmulatedTestUtilities.failNow());
-	}
-	
-	/**
-	 * Either passes or skips.
-	 *
-	 * @param __millis The time to finish.
-	 * @param __pass Does the test pass?
-	 * @return The event.
-	 * @since 2020/03/06
-	 */
-	public static TestCompleteEvent passOrSkipAt(long __millis, boolean __pass)
-	{
-		return (__pass ? EmulatedTestUtilities.passAt(__millis) :
-			EmulatedTestUtilities.skipAt(__millis));
-	}
-	
-	/**
-	 * Passes, skips, or fails at the given time.
-	 * 
-	 * @param __millis The time to finish.
-	 * @param __exitCode The status code.
-	 * @return The completion event.
-	 * @since 2020/06/22
-	 */
-	public static TestCompleteEvent passSkipOrFailAt(long __millis,
-		int __exitCode)
-	{
-		if (__exitCode == ExitValueConstants.SUCCESS)
-			return EmulatedTestUtilities.passAt(__millis);
-		else if (__exitCode == ExitValueConstants.SKIPPED)
-			return EmulatedTestUtilities.skipAt(__millis);
-		return EmulatedTestUtilities.failAt(__millis);
-	}
-	
-	/**
-	 * Depending on the exit code, this will pass, fail, or skip a test.
-	 *
-	 * @param __exitCode The exit code.
-	 * @return The completion event.
-	 * @since 2020/03/07
-	 */
-	public static TestCompleteEvent passSkipOrFailNow(int __exitCode)
-	{
-		return EmulatedTestUtilities.passSkipOrFailAt(
-			System.currentTimeMillis(), __exitCode);
-	}
-	
-	/**
-	 * Reads test services from the JAR. 
-	 *
-	 * @param __path The path to read from.
-	 * @throws IOException On read failures.
-	 * @since 2020/06/22
-	 */
-	public static Collection<String> readJarServices(Path __path)
-		throws IOException, NullPointerException
-	{
-		if (__path == null)
-			throw new NullPointerException("NARG");
-		
-		try (InputStream in = Files.newInputStream(
-			__path, StandardOpenOption.READ);
-			ZipInputStream zip = new ZipInputStream(in))
-		{
-			return EmulatedTestUtilities.readJarServices(zip);
-		}
-	}
-	
-	/**
-	 * Reads test services from the JAR. 
-	 *
-	 * @param __zip The ZIP to read.
-	 * @throws IOException On read failures.
-	 * @since 2020/05/25
-	 */
-	public static Collection<String> readJarServices(ZipInputStream __zip)
-		throws IOException
-	{
-		// Resultant collection
-		Collection<String> result = new LinkedHashSet<>();
-		
-		// Potential resource file data
-		byte[] resourceData = null;
-		
-		// Go through entries within the ZIP
-		ZipEntry entry;
-		while (null != (entry = __zip.getNextEntry()))
-		{
-			// Only consider the services file
-			if (!Objects.equals(EmulatedTestExecutor.SERVICE_RESOURCE,
-				entry.getName()))
-				continue;
-			
-			// Copy file data
-			try (ByteArrayOutputStream baos = new ByteArrayOutputStream())
-			{
-				// Copy bytes
-				byte[] buf = new byte[4096];
-				for (;;)
-				{
-					int rc = __zip.read(buf);
-					
-					if (rc < 0)
-						break;
-					
-					baos.write(buf, 0, rc);
-				}
-				
-				// Set resource data and stop finding entries
-				resourceData = baos.toByteArray();
-				break;
-			}
-		}
-		
-		// Read in service list
-		if (resourceData != null)
-			try (BufferedReader br = new BufferedReader(new InputStreamReader(
-				new ByteArrayInputStream(resourceData),
-					StandardCharsets.UTF_8)))
-			{
-				for (;;)
-				{
-					// End of file?
-					String ln = br.readLine();
-					if (ln == null)
-						break;
-					
-					// Only consider lines with content
-					ln = ln.trim();
-					if (!ln.isEmpty())
-						result.add(ln);
-				}
-			}
-		
-		return result;
-	}
-	
-	/**
-	 * Skips at the given time.
-	 * 
-	 * @param __millis The time to skip at.
-	 * @return The completion event.
-	 * @since 2020/06/22
-	 */
-	private static TestCompleteEvent skipAt(long __millis)
-	{
-		return new TestCompleteEvent(__millis,
-			TestResult.ResultType.SKIPPED);
-	}
-	
-	/**
-	 * Returns a skipping completion.
-	 *
-	 * @return An event.
-	 * @since 2020/03/06
-	 */
-	public static TestCompleteEvent skipNow()
-	{
-		return EmulatedTestUtilities.skipAt(System.currentTimeMillis());
-	}
-	
-	/**
-	 * Returns a test starting now.
-	 *
-	 * @param __millis The starting time.
-	 * @return An event
-	 * @since 2020/06/22
-	 */
-	public static TestStartEvent startAt(long __millis)
-	{
-		return EmulatedTestUtilities.startAt(__millis, null);
-	}
-	
-	/**
-	 * Returns a test starting now.
-	 *
-	 * @param __test The test information.
-	 * @return An event
-	 * @since 2020/06/22
-	 */
-	public static TestStartEvent startAt(long __millis,
-		TestDescriptorInternal __test)
-	{
-		return new TestStartEvent(__millis,
-			(__test == null ? null : __test.getId()));
-	}
-	
-	/**
-	 * Returns a test starting now.
-	 *
-	 * @return An event
-	 * @since 2020/03/06
-	 */
-	public static TestStartEvent startNow()
-	{
-		return EmulatedTestUtilities.startNow(null);
-	}
-	
-	/**
-	 * Returns a test starting now.
-	 *
-	 * @param __test The test information.
-	 * @return An event
-	 * @since 2020/03/06
-	 */
-	public static TestStartEvent startNow(TestDescriptorInternal __test)
-	{
-		return EmulatedTestUtilities.startAt(
-			System.currentTimeMillis(), __test);
-	}
-}
